Minor fixes. (#62)
* VkPhysicalDeviceVulkan12Features::uniformAndStorageBuffer8BitAccess is not required.

* EXT suffix required for vkCmdSetCullMode.
Since VK_EXT_extended_dynamic_state is core in Vulkan 1.2 and the application is targeting to Vulkan 1.2, vkCmdSetCullMode function pointer is not loaded.

* Fix wrong ImGui_ImplVulkan_RemoveTexture usage.

* Fix destroying texture that is currently referred by ImGui descriptor set.
When asset is changed while ImGui is showing any texture, the frame will record the command buffer with the invalidated descriptor set. To avoid this, the invalidated descriptor sets have to be updated with the currently alive texture (fallbackTexture is used in the change).